requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.0
uvicorn==0.30.6
starlette==0.38.6
pydantic==2.9.2
pydantic_core==2.23.4

# HTTP and Web Requests
httpx==0.28.1
httpcore==1.0.7
requests==2.32.3
aiohttp==3.12.15
aiofiles==24.1.0

# AI and OpenAI
openai==1.70.0

# Twilio for Voice Calls
twilio==9.3.2

# Email Services
yagmail==0.15.293
email-validator==2.1.0.post1

# Data Processing
pandas==2.2.3
numpy==2.2.4
openpyxl==3.1.5

# Document Processing
pdfminer.six==20231228
python-docx==1.1.0
PyPDF2==3.0.1
lxml==5.4.0

# Database
sqlalchemy

# Environment and Configuration
python-dotenv==1.0.1

# CORS and Security
flask-cors==4.0.0

# WebSocket Support
websockets==13.1

# Utilities
click==8.1.7
python-dateutil==2.9.0.post1
pytz==2025.2
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
urllib3==2.2.3

# Optional: Streamlit for alternative UI
streamlit

# Optional: Flask for additional web services
Flask==3.0.0
Werkzeug==3.0.1